Clean up current course components
